working/requirements.txt
Flask==2.0.2
pyttsx3==2.90
SpeechRecognition==3.8.1
google-generativeai==0.2.0
pyautogui==0.9.53
comtypes==1.1.10
subprocess
